refactor: moving new notification logging to websocketBroadcast
